refactor(desktop): move HighlighterWorker to own module
